change done on 20/04

+++ libft/ft_put_unbr_fd.c
@@ -6,13 +6,13 @@
 /*   By: lagea <lagea@student.s19.be>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/21 21:34:11 by lagea             #+#    #+#             */
-/*   Updated: 2024/04/20 15:39:42 by lagea            ###   ########.fr       */
+/*   Updated: 2024/04/20 20:04:36 by lagea            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static void	print_nb(long n, int fd)
+static void	print_nb(unsigned int n, int fd)
 {
 	if (n / 10)
 	{
